requirements.txt
sentence-transformers
faiss-cpu
transformers
pypdf
numpy